feed table config
