+++ src/interfaces/feed-list.ts
@@ -1,0 +1,12 @@
+type FeedList = {
+  url: string,
+  id: string | undefined,
+  title: string
+  published: string,
+  keywords: string[],
+  score: number,
+  matchedCases: number,
+  matchedBlogs: number,
+};
+
+export default FeedList;
